requirements.txt
asgiref==3.8.1
Django==5.1.3
django-appconf==1.0.6
django-environ==0.11.2
django-imagekit==5.0.0
django-widget-tweaks==1.5.0
djangorestframework==3.15.2
gunicorn==23.0.0
packaging==24.2
pilkit==3.0
pillow==11.0.0
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.2
